% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{tcga_gistic_available}
\alias{tcga_gistic_available}
\title{List Available Datasets}
\usage{
tcga_gistic_available()
}
\value{
a dataframe listing available datasets
}
\description{
List Available Datasets
}
\examples{
tcga_gistic_available()

}
