% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{tcga_gistic_load}
\alias{tcga_gistic_load}
\title{Load TCGA GISTIC data}
\usage{
tcga_gistic_load(
  cohort,
  source = "Firehose",
  cnLevel = c("all", "deep", "shallow"),
  verbose = TRUE
)
}
\arguments{
\item{cohort}{abbreviation of TCGA cohort See \code{\link[=tcga_gistic_available]{tcga_gistic_available()}} for valid values (string)}

\item{source}{source of the data (currently we only support 'Firehose' data (string)}

\item{cnLevel}{level of CN changes to use. Can be 'all', 'deep' or 'shallow'. Default uses all i.e, genes with both 'shallow' or 'deep' CN changes}

\item{verbose}{verbosity (flag)}
}
\value{
A maftools & CRUX-compatible list of summarized data.
}
\description{
Load TCGA GISTIC objects into R. Streams data from \href{https://github.com/CCICB/TCGAgisticDB}{TCGAgisticDB} repo.
}
\examples{
# Load libraries
library(TCGAgistic)
library(maftools)

# Load dataset
gistic <- tcga_gistic_load("ACC", source = "Firehose", cnLevel = "all")

# Visualise results
gisticChromPlot(gistic)
}
